requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==11.1.0
google-cloud-storage==2.17.0
pydantic-settings==2.3.1
python-multipart==0.0.9